Convert Variance-Covariance Stored in Tibble to a List of matrices

Storing the variance and covariances in the tibble read from the input file in a list of matrices.

build_matrix(ps_input_tibble)

Arguments

ps_input_tibble
tibble containing all variance-covariance components

Details

--- title: Constructing multivariate estimates from analyses by parts in a matrix date: "`r Sys.Date()`" ---

Examples

sInputFile <- system.file("extdata","VCE_results.csv", package = "rvcetools")
tbl_vce <- read_vce(ps_input_file = sInputFile)

#> Parsed with column specification:
#> cols(
#>   type = col_character(),
#>   traits = col_character(),
#>   random_effect = col_character(),
#>   estimate = col_character(),
#>   STD_ERR_estimate = col_character(),
#>   model_name = col_character(),
#>   data_subset_number = col_character(),
#>   used_seed = col_character(),
#>   trait_combination = col_character()
#> )

l_mat <- build_matrix(ps_input_tibble = tbl_vce)